Minimum code refactor at SourceAdapter, leaving the viewHolder object into a separate class in order to increase readability
